requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
httpx>=0.27
pydantic>=2
python-dotenv>=1.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt<4.1
python-multipart>=0.0.9
